+++ administrator/urls.py
@@ -4,8 +4,7 @@ from .views import ParkingAdminCreate, CreateCarAdminView, SettingAdminInfo, Dow
 
 app_name = 'administrator'
 urlpatterns = [
-    path('', views.check_superuser, name='check_superuser'),
-    path('index/', views.index, name='index'),
+    path('', views.index, name='index'),
     path('ajax/', views.test_ajax_response),
     path('create/', ParkingAdminCreate.as_view(), name='create'),
     path('admin_main/', views.admin_main, name='admin_main'),
